+++ packages/admin-stories/src/admin/stack/RouterTabsNestedStack.tsx
@@ -1,4 +1,4 @@
-import { RouterTab, RouterTabs, Stack, StackPage, StackSwitch, StackSwitchApiContext } from "@comet/admin";
+import { Breadcrumbs, RouterTab, RouterTabs, Stack, StackPage, StackSwitch, StackSwitchApiContext } from "@comet/admin";
 import { Button } from "@material-ui/core";
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
@@ -7,9 +7,11 @@ import StoryRouter from "storybook-react-router";
 function Story() {
     return (
         <Stack topLevelTitle="Root Stack">
+            <Breadcrumbs />
             <RouterTabs>
                 <RouterTab label="Page 1" path="">
                     <Stack topLevelTitle="Nested Stack">
+                        <Breadcrumbs />
                         <StackSwitch>
                             <StackPage name="table">
                                 <StackSwitchApiContext.Consumer>