+++ packages/admin/src/mui/breadcrumbs/index.ts
@@ -1,1 +1,2 @@
 export * from "./Breadcrumbs";
+export { CometAdminBreadcrumbsClassKeys, CometAdminBreadcrumbsThemeProps } from "./Breadcrumbs.styles";